NOT WORKING ft_printptr. multiples errors when compiling, execute test and debug

+++ 42projects/ft_printf/inc/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: odruke <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/30 12:16:00 by odruke            #+#    #+#             */
-/*   Updated: 2023/05/07 22:09:05 by druke            ###   ########.fr       */
+/*   Updated: 2023/05/08 21:06:10 by odruke           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,8 @@
 # include "stdarg.h"
 # include <stdio.h>
 # include <unistd.h>
-int	ft_printptr(int *ptr);
+# include <stdint.h>
+int	ft_printptr(unsigned long long *ptr);
 int	ft_printnbr(int i);
 int	ft_printf(char *str, ...);
 int	ft_printchar(int  c);
